+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+Application.ensure_started(:telemetry)
 ExUnit.start()